Major Changes
Added location prototypes
Instantiated an instance of Location() for each location
Stored the location instances in a global array
Wrote an event handler to display each location using the current
location and the (global) array
of location instances
currently working on only adding score the first time a location is gone
to. Will most likely use an array
Also trying to figure out how to draw a map that doesn't look like a
pre-school student did it.